Update PR with comments from video meeting
